+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 ### master
 
+ - Fixed an issue where "Derived" operations could not be run in non-local mode.
+ - The Python API now allows running external computations using the `@external` decorator.
+ - Upgraded to Apache Spark 2.3.2.
+
 ### 2.5.0
 
  - Cleaner works again! Just pick _"Cleaner"_ from the user menu as an administrator and click